requirements.txt
# Core dependencies
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Data processing
numpy>=1.26.0
pandas>=2.2.0

# Machine Learning
scikit-learn>=1.5.0
xgboost>=2.1.0

# Embeddings
sentence-transformers>=3.0.0
torch>=2.0.0  # For sentence-transformers

# Visualization
umap-learn>=0.5.5
matplotlib>=3.9.0
plotly>=5.24.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0
structlog>=24.4.0
rich>=13.7.0

# Testing
pytest>=8.3.0
pytest-cov>=5.0.0
pytest-asyncio>=0.24.0
hypothesis>=6.112.0

# Code quality
black>=24.8.0
ruff>=0.6.0
mypy>=1.11.0

# Development
ipython>=8.27.0
jupyter>=1.1.0

# Streamlit UI
streamlit>=1.40.0
